moved expression interfaces and guards to its own files.
