+++ src/editor/components/operators/QueryEditorBoolOperator.tsx
@@ -1,18 +1,14 @@
 import React, { useCallback } from 'react';
 import { Select } from '@grafana/ui';
-import { QueryEditorOperatorExpression } from '../types';
-import { QueryEditorOperatorDefinition } from '../../types';
 import { SelectableValue } from '@grafana/data';
+import { QueryEditorOperatorDefinition } from '../../types';
 import { QueryEditorExpressionType } from '../../../types';
+import { QueryEditorBoolOperatorExpression } from '../../expressions';
 
 interface Props {
   value: boolean | undefined;
   onChange: (expression: QueryEditorBoolOperatorExpression) => void;
   operator: QueryEditorOperatorDefinition;
-}
-
-export interface QueryEditorBoolOperatorExpression extends QueryEditorOperatorExpression {
-  value: boolean;
 }
 
 const options: Array<SelectableValue<boolean>> = [
@@ -37,10 +33,4 @@ export const QueryEditorBoolOperator: React.FC<Props> = props => {
   );
 
   return <Select width={15} options={options} value={props.value} onChange={onChange} menuPlacement="bottom" />;
-};
-
-export const isBoolOperator = (
-  expression: QueryEditorOperatorExpression | undefined
-): expression is QueryEditorBoolOperatorExpression => {
-  return typeof (expression as QueryEditorBoolOperatorExpression)?.value === 'boolean';
 };
